+++ packages/mobile/src/alpha/select/DefaultSelectControl.tsx
@@ -63,7 +63,6 @@ export const DefaultSelectControlComponent = memo(
         focusedBorderWidth = bordered ? undefined : 200,
         maxSelectedOptionsToShow = 3,
         accessibilityLabel,
-        accessibilityHint,
         hiddenSelectedOptionsLabel = 'more',
         removeSelectedOptionAccessibilityLabel = 'Remove',
         style,
@@ -124,6 +123,37 @@ export const DefaultSelectControlComponent = memo(
         });
         return map;
       }, [options]);
+
+      const singleValueContent = useMemo(() => {
+        const option = !isMultiSelect ? optionsMap.get(value as SelectOptionValue) : undefined;
+        const label = option?.label ?? option?.description ?? option?.value ?? placeholder;
+        return hasValue ? label : placeholder;
+      }, [hasValue, isMultiSelect, optionsMap, placeholder, value]);
+
+      const computedControlAccessibilityLabel = useMemo(() => {
+        // For multi-select, set the label to the content of each selected value and the hidden selected options label
+        if (isMultiSelect) {
+          const selectedValues = (value as SelectOptionValue[])
+            .map((v) => {
+              const option = optionsMap.get(v);
+              return option?.label ?? option?.description ?? option?.value ?? v;
+            })
+            .slice(0, maxSelectedOptionsToShow)
+            .join(', ');
+          return `${accessibilityLabel}, ${(value as SelectOptionValue[]).length > 0 ? selectedValues : (placeholder ?? '')}${(value as SelectOptionValue[]).length > maxSelectedOptionsToShow ? ', ' + hiddenSelectedOptionsLabel : ''}`;
+        }
+        // If value is React node, fallback to only using passed in accessibility label
+        return `${accessibilityLabel ?? ''}${typeof singleValueContent === 'string' ? ', ' + singleValueContent : ''}`;
+      }, [
+        accessibilityLabel,
+        hiddenSelectedOptionsLabel,
+        isMultiSelect,
+        maxSelectedOptionsToShow,
+        optionsMap,
+        placeholder,
+        singleValueContent,
+        value,
+      ]);
 
       // Prop value doesn't have default value because it affects the color of the
       // animated caret
@@ -220,24 +250,21 @@ export const DefaultSelectControlComponent = memo(
           );
         }
 
-        const option = !isMultiSelect ? optionsMap.get(value as SelectOptionValue) : undefined;
-        const label = option?.label ?? option?.description ?? option?.value ?? placeholder;
-        const content = hasValue ? label : placeholder;
-        return typeof content === 'string' ? (
+        return typeof singleValueContent === 'string' ? (
           <Text color={hasValue ? 'fg' : 'fgMuted'} ellipsize="tail" font="body" textAlign="left">
-            {content}
+            {singleValueContent}
           </Text>
         ) : (
-          content
+          singleValueContent
         );
       }, [
         hasValue,
         isMultiSelect,
-        optionsMap,
-        placeholder,
+        singleValueContent,
         value,
         maxSelectedOptionsToShow,
         hiddenSelectedOptionsLabel,
+        optionsMap,
         removeSelectedOptionAccessibilityLabel,
         disabled,
         onChange,
@@ -247,8 +274,7 @@ export const DefaultSelectControlComponent = memo(
         () => (
           <TouchableOpacity
             ref={ref}
-            accessibilityHint={accessibilityHint}
-            accessibilityLabel={accessibilityLabel}
+            accessibilityLabel={computedControlAccessibilityLabel}
             accessibilityRole="button"
             disabled={disabled}
             onPress={() => setOpen((s) => !s)}
@@ -293,8 +319,7 @@ export const DefaultSelectControlComponent = memo(
         ),
         [
           ref,
-          accessibilityHint,
-          accessibilityLabel,
+          computedControlAccessibilityLabel,
           disabled,
           styles?.controlInputNode,
           styles?.controlStartNode,
@@ -313,7 +338,11 @@ export const DefaultSelectControlComponent = memo(
 
       const endNode = useMemo(
         () => (
-          <Pressable disabled={disabled} onPress={() => setOpen((s) => !s)}>
+          <Pressable
+            accessible={customEndNode ? true : false}
+            disabled={disabled}
+            onPress={() => setOpen((s) => !s)}
+          >
             <HStack
               alignItems="center"
               flexGrow={1}