+++ packages/web/CHANGELOG.md
@@ -8,6 +8,30 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.38.5 (1/23/2026 PST)
+
+#### 🐞 Fixes
+
+- Improve keyboard navigation and ARIA labels on Select and Combobox. [[#250](https://github.com/coinbase/cds/pull/250)]
+
+## 8.38.4 (1/22/2026 PST)
+
+#### 🐞 Fixes
+
+- Fixed spacing props not working on web Button.
+
+## 8.38.3 (1/22/2026 PST)
+
+#### 🐞 Fixes
+
+- Fix: destructure unused props from default horizontal stepper components to prevent dev mode React warnings. [[#324](https://github.com/coinbase/cds/pull/324)]
+
+## 8.38.2 (1/22/2026 PST)
+
+#### 🐞 Fixes
+
+- FocusTrap supports single focusable child and updates to its tests. [[#306](https://github.com/coinbase/cds/pull/306)]
+
 ## 8.38.1 (1/15/2026 PST)
 
 #### 🐞 Fixes